requirements.txt
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
python-dotenv==1.0.1
drf-spectacular==0.27.2